задание 4.
